requirements.txt
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
Django==5.1.5
django-ckeditor==6.7.2
django-js-asset==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
googletrans==4.0.2
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
redis==5.2.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
